王晓松 完成Pojo生成实体类功能

+++ leno-dev/src/main/java/com/yogo/agent/service/ConfService.java
@@ -3,15 +3,21 @@ package com.yogo.agent.service;
 import com.alibaba.fastjson.JSON;
 import com.yogo.agent.common.back.ResultInfo;
 import com.yogo.agent.common.enums.ResultEnum;
+import com.yogo.agent.common.utils.leno.util.LenoDBOperation;
 import com.yogo.agent.entity.ConfEntity;
 import com.yogo.agent.mapper.ConfMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @Author owen
@@ -20,7 +26,7 @@ import java.util.List;
  **/
 @Slf4j
 @Service
-public class ConfService {
+public class ConfService extends LenoDBOperation {
 
     private final ConfMapper confMapper;
 
@@ -30,7 +36,12 @@ public class ConfService {
     }
 
     public ResultInfo addConf(ConfEntity conf, Principal principal) {
-        conf.setUser(principal.getName());
+        String name = principal.getName();
+        List<ConfEntity> list = confMapper.selByUser(name);
+        if (list.isEmpty()) {
+            conf.setEnable(1);
+        }
+        conf.setUser(name);
         confMapper.insertSelective(conf);
         return ResultInfo.response(ResultEnum.SAVE_OK);
     }
@@ -39,6 +50,19 @@ public class ConfService {
         conf.setUser(principal.getName());
         confMapper.updateByPrimaryKeySelective(conf);
         return ResultInfo.response(ResultEnum.EDIT_OK);
+    }
+
+    public ResultInfo testConf(ConfEntity conf) {
+        try {
+            synchronized (conf.getId()) {
+                String url = "jdbc:mysql://" + conf.getUrl() + "/" + conf.getLib() + "?useUnicode=true&characterEncoding=utf-8&useSSL=false";
+                getConnection(url, conf.getUsername(), conf.getPassword());
+            }
+            return ResultInfo.response(ResultEnum.CONN_OK);
+        } catch (Exception e) {
+            log.info("连接失败" + e.getMessage());
+            return ResultInfo.response(ResultEnum.CONN_ERROR);
+        }
     }
 
     public ResultInfo selConf(Principal principal) {
@@ -59,9 +83,34 @@ public class ConfService {
     }
 
     public ResultInfo delConf(Long id) {
+        ConfEntity config = confMapper.selectByPrimaryKey(id);
+        Integer status = config.getEnable();
+        if (status == 1) {
+            return ResultInfo.response(ResultEnum.CAN_NOT_DELETE);
+        }
         confMapper.deleteByPrimaryKey(id);
         return ResultInfo.response(ResultEnum.DEL_OK);
     }
 
 
+    public List<String> getTable(ConfEntity conf) {
+        if (Objects.isNull(conf)) {
+            return new ArrayList<>();
+        }
+        ArrayList<String> list = new ArrayList<>();
+        synchronized (conf.getId()) {
+            String url = "jdbc:mysql://" + conf.getUrl() + "/" + conf.getLib() + "?useUnicode=true&characterEncoding=utf-8&useSSL=false";
+            try (Connection conn = getConnection(url, conf.getUsername(), conf.getPassword())) {
+                ResultSet rs = conn.getMetaData().getTables(conn.getCatalog(), conf.getLib(), "%", new String[]{"TABLE"});
+                while (rs.next()) {
+                    String table = rs.getString("TABLE_NAME");
+                    list.add(table);
+                }
+                return list;
+            } catch (Exception e) {
+                log.error("getTableMethodException :" + e.getMessage());
+                return new ArrayList<>();
+            }
+        }
+    }
 }
